move metadata.rs to new cluster mod
