+++ scylla/src/session/session_test.rs
@@ -1,5 +1,7 @@
 use crate as scylla;
 use crate::batch::{Batch, BatchStatement};
+use crate::cluster::metadata::Strategy::NetworkTopologyStrategy;
+use crate::cluster::metadata::{CollectionType, ColumnKind, CqlType, NativeType, UserDefinedType};
 use crate::frame::response::result::Row;
 use crate::frame::value::ValueList;
 use crate::prepared_statement::PreparedStatement;
@@ -10,10 +12,6 @@ use crate::statement::Consistency;
 use crate::tracing::TracingInfo;
 use crate::transport::cluster::Datacenter;
 use crate::transport::errors::{BadKeyspaceName, BadQuery, DbError, QueryError};
-use crate::transport::metadata::Strategy::NetworkTopologyStrategy;
-use crate::transport::metadata::{
-    CollectionType, ColumnKind, CqlType, NativeType, UserDefinedType,
-};
 use crate::transport::partitioner::{
     calculate_token_for_partition_key, Murmur3Partitioner, Partitioner, PartitionerName,
 };
