+++ scylla/src/transport/locator/test.rs
@@ -3,12 +3,10 @@ use rand_chacha::ChaCha8Rng;
 use uuid::Uuid;
 
 use super::{ReplicaLocator, ReplicaSet};
+use crate::cluster::metadata::{Keyspace, Metadata, Peer, Strategy};
 use crate::connection::PoolConfig;
 use crate::sharding::Token;
-use crate::transport::{
-    metadata::{Keyspace, Metadata, Peer, Strategy},
-    Node,
-};
+use crate::transport::Node;
 use crate::transport::{NodeAddr, NodeRef};
 
 use std::collections::HashSet;
