+++ scylla/src/transport/node.rs
@@ -21,7 +21,7 @@ use std::{
     },
 };
 
-use super::metadata::{PeerEndpoint, UntranslatedEndpoint};
+use crate::cluster::metadata::{PeerEndpoint, UntranslatedEndpoint};
 
 /// This enum is introduced to support address translation only upon opening a connection,
 /// as well as to cope with a bug present in older Cassandra and Scylla releases.
